# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  apierror.ErrorResponse:
    properties:
      error:
        description: Stable error code (e.g., "invalid_credentials")
        type: string
      message:
        description: Human-readable message
        type: string
    type: object
...
  dto.GitIntegrationCallbackRequest:
    properties:
      code:
        type: string
      error:
        description: Error — если провайдер вернул ?error=... (например access_denied).
        type: string
      state:
        type: string
    required:
    - state
    type: object
  dto.GitIntegrationCallbackResponse:
    properties:
      provider:
        type: string
      status:
        $ref: '#/definitions/dto.GitIntegrationStatusResponse'
    type: object
  dto.GitIntegrationInitRequest:
    properties:
      byo_client_id:
        description: ByoClientID / ByoClientSecret — только для self-hosted GitLab.
        type: string
      byo_client_secret:
        type: string
      host:
        description: Host — только для self-hosted GitLab. Пусто = shared.
        type: string
      redirect_uri:
        description: |-
          RedirectURI — куда провайдер вернёт пользователя после consent. Должен совпадать
          с client_secret-конфигом OAuth App у провайдера.
        type: string
    required:
    - redirect_uri
    type: object
  dto.GitIntegrationInitResponse:
    properties:
      authorize_url:
        type: string
      state:
        type: string
    type: object
  dto.GitIntegrationRevokeResponse:
    properties:
      provider:
        type: string
      remote_revoke_failed:
        type: boolean
    type: object
  dto.GitIntegrationStatusResponse:
    properties:
      account_login:
        type: string
      connected:
        type: boolean
      connected_at:
        type: string
      expires_at:
        type: string
      host:
        type: string
      id:
        type: string
      provider:
        type: string
      scopes:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Backend API с авторизацией на JWT токенах
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend API
  version: "1.0"
paths:
...
  /integrations/accounts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.GitIntegrationStatusResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Список подключённых OAuth-аккаунтов
      tags:
      - git-integrations
  /integrations/accounts/{id}:
    delete:
      parameters:
      - description: Credential ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GitIntegrationRevokeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Отозвать подключённый OAuth-аккаунт по id
      tags:
      - git-integrations
  /integrations/github/auth/callback:
    post:
      consumes:
      - application/json
      parameters:
      - description: code + state
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GitIntegrationCallbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GitIntegrationCallbackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "410":
          description: user_cancelled / invalid_state
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "502":
          description: provider_unreachable
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Завершение GitHub OAuth (code + state от провайдера)
      tags:
      - git-integrations
  /integrations/github/auth/init:
    post:
      consumes:
      - application/json
      description: Возвращает URL, куда фронт открывает попап/таб для авторизации.
      parameters:
      - description: redirect_uri
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GitIntegrationInitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GitIntegrationInitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Старт GitHub OAuth
      tags:
      - git-integrations
  /integrations/github/auth/revoke:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GitIntegrationRevokeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Отозвать подключение GitHub
      tags:
      - git-integrations
  /integrations/github/auth/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GitIntegrationStatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Статус подключения GitHub
      tags:
      - git-integrations
  /integrations/gitlab/auth/callback:
    post:
      consumes:
      - application/json
      parameters:
      - description: code + state
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GitIntegrationCallbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GitIntegrationCallbackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "410":
          description: user_cancelled / invalid_state
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "502":
          description: provider_unreachable
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Завершение GitLab OAuth (code + state от провайдера)
      tags:
      - git-integrations
  /integrations/gitlab/auth/init:
    post:
      consumes:
      - application/json
      description: Если в теле передан host+byo_client_id+byo_client_secret — self-hosted;
        иначе gitlab.com.
      parameters:
      - description: redirect_uri (+ host/byo_* для self-hosted)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GitIntegrationInitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GitIntegrationInitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Старт GitLab OAuth (gitlab.com shared или self-hosted BYO)
      tags:
      - git-integrations
  /integrations/gitlab/auth/revoke:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GitIntegrationRevokeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Отозвать подключение GitLab
      tags:
      - git-integrations
  /integrations/gitlab/auth/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GitIntegrationStatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Статус подключения GitLab
      tags:
      - git-integrations
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Long-lived API key for programmatic access. Format: wibe_<key>'
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  OAuth2Password:
    flow: password
    tokenUrl: /api/v1/auth/login
    type: oauth2
swagger: "2.0"